passed simple kernel

+++ _simple.h
@@ -9,7 +9,7 @@ __global__ void kernel_simple(float *a, float *x, float *y, int XR, int XC, int 
   int r = by*BY + ty;
   int c = bx*BX + tx;
   
-  int s = 0;
+  float s = 0;
   for (int i=0; i<XC; i++)
     s += GET2D(x, r, i, XC) * GET2D(y, i, c, YC);
   GET2D(a, r, c, YC) = s;
